requirements.txt
numpy==1.24.3
scikit-learn==1.3.0
pathlib==1.0.1
pandas==2.0.3
scipy==1.10.1
joblib==1.3.1
markdown==3.3.0
jinja2==3.0.0
nbformat==5.1.0
matplotlib==3.4.0
networkx==2.6.0
pyyaml==5.4.0
requests==2.26.0
beautifulsoup4==4.9.0
python-markdown-math==0.8
python-markdown-mathjax==0.1
markdown-include==0.6.0
markdown-checklist==0.1.0
markdown-fenced-code-tabs==0.1.0
markdown-mermaid==0.1.0
markdown-pdf==0.0.1
markdown-pdf-include==0.0.1
markdown-pdf-mermaid==0.0.1
markdown-pdf-mathjax==0.0.1
markdown-pdf-checklist==0.0.1
markdown-pdf-fenced-code-tabs==0.0.1
markdown-pdf-include==0.0.1
markdown-pdf-mermaid==0.0.1
markdown-pdf-mathjax==0.0.1
markdown-pdf-checklist==0.0.1
markdown-pdf-fenced-code-tabs==0.0.1 